Day six

+++ ex012.py
@@ -1,7 +1,8 @@
-p = int(input('Digite o valor do produto:'))
+p = float(input('Digite o valor do produto:'))
 d = p * 0.05
 vf = p - d
 print(f'O preço do produto e de R$ {p} com o desconto de 5% seu preço vai ser de R$ {vf:.2f}')
+
